chore(swingset): move gc.js to gc-and-finalize.js
get it out of the way for other kernel GC things to use the filenames
